fix: enum validation to disallow reserved names
