+++ engine/baml-lib/baml/tests/validation_files/enum/enum_is_valid_without_generator.baml
@@ -1,0 +1,5 @@
+enum Test {
+    None
+    Hello
+    World
+}
